orderbook-backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0